+++ src/main/java/com/ldtteam/blockui/views/ScrollingList.java
@@ -2,7 +2,8 @@ package com.ldtteam.blockui.views;
 
 import com.ldtteam.blockui.Pane;
 import com.ldtteam.blockui.PaneParams;
-import com.ldtteam.blockui.support.DataProviders.DataProvider;
+import com.ldtteam.blockui.util.records.SizeI;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.List;
 import java.util.function.IntSupplier;
@@ -119,6 +120,39 @@ public class ScrollingList extends ScrollingView
     public int getListElementIndexByPane(final Pane pane)
     {
         return ((ScrollingListContainer) container).getListElementIndexByPane(pane);
+    }
+
+    /**
+     * Interface for a data provider that updates pane scrolling list pane info.
+     */
+    public interface DataProvider
+    {
+        /**
+         * Override this to provide the number of rows.
+         *
+         * @return number of rows in the list
+         */
+        int getElementCount();
+
+        /**
+         * Override this to pick a custom size for this element. Tuple arguments are width and height, in that order.
+         *
+         * @param index   the index of the row/list element
+         * @param rowPane the parent Pane for the row, containing the elements to update
+         * @return a new size for the element, or null to use the template element size.
+         */
+        default @Nullable SizeI getElementSize(int index, Pane rowPane)
+        {
+            return null;
+        }
+
+        /**
+         * Override this to update the Panes for a given row.
+         *
+         * @param index   the index of the row/list element
+         * @param rowPane the parent Pane for the row, containing the elements to update
+         */
+        void updateElement(int index, Pane rowPane);
     }
 
     @FunctionalInterface
